+++ xmlschema/tests/test_all.py
@@ -40,5 +40,5 @@ if __name__ == '__main__':
 
     testfiles = get_testfiles(os.path.dirname(os.path.abspath(__file__)))
     globals().update(tests_factory(make_test_schema_function, testfiles, 'schema', 'xsd'))
-    globals().update(tests_factory(make_decoder_test_function, testfiles, 'decoding', 'xml'))
+    globals().update(tests_factory(make_decoder_test_function, testfiles, 'decoder', 'xml'))
     unittest.main()